+++ projects/plugins/jetpack/extensions/store/wordpress-com/actions.ts
@@ -1,0 +1,30 @@
+/**
+ * Types
+ */
+import type { Plan } from './types';
+import type { AIFeatureProps } from './types';
+
+const actions = {
+	setPlans( plans: Array< Plan > ) {
+		return {
+			type: 'SET_PLANS',
+			plans,
+		};
+	},
+
+	fetchFromAPI( url: string ) {
+		return {
+			type: 'FETCH_FROM_API',
+			url,
+		};
+	},
+
+	storeAiAssistantFeature( feature: AIFeatureProps ) {
+		return {
+			type: 'STORE_AI_ASSISTANT_FEATURE',
+			feature,
+		};
+	},
+};
+
+export default actions;
